+++ client/src/Root.tsx
@@ -1,4 +1,4 @@
-import { Outlet, useNavigate } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 import Header from './components/header/Header';
 import Toast from './components/toast/toast';
 import { useEffect } from 'react';
@@ -9,7 +9,6 @@ import { storage } from './common/utils/storage';
 
 function Root() {
   const dispatch = useAppDispatch();
-  const navigate = useNavigate();
 
   const getUser = async () => {
     const userInfo = await getUserInfo();
@@ -21,7 +20,6 @@ function Root() {
     const memberId = storage.get('memberId');
     if (accessToken && memberId) {
       getUser();
-      navigate('/');
     }
   }, []);
