+++ client/src/components/route/PublicRoute.tsx
@@ -1,0 +1,24 @@
+import { useMatch, useNavigate } from 'react-router-dom';
+import { ReactNode, useEffect } from 'react';
+import { storage } from '@/common/utils/storage';
+
+type Props = {
+  children: ReactNode;
+};
+
+const PublicRoute = ({ children }: Props) => {
+  const navigate = useNavigate();
+  const isLoginPage = useMatch('/users/login');
+  const isSignUpPage = useMatch('/users/signup');
+
+  useEffect(() => {
+    const accessToken = storage.get('accessToken');
+    if (accessToken && (isSignUpPage || isLoginPage)) {
+      navigate('/');
+    }
+  }, []);
+
+  return <>{children}</>;
+};
+
+export default PublicRoute;
